requirements.txt
numpy
pandas
Pillow
imageio
scipy
opencv-python
matplotlib
tensorflow
torch
imageio
dlib
imutils